[tests] add build tests for BR features (#6570)
BR features are tested with "Border Router" CI task but it is built
with gcc by default. This commit adds BR features to posix build
checks which apply clang/clang++.

+++ src/posix/platform/infra_if.cpp
@@ -295,7 +295,7 @@ uint32_t InfraNetif::GetFlags(void) const
 
     close(sock);
 
-    return ifReq.ifr_flags;
+    return static_cast<uint32_t>(ifReq.ifr_flags);
 }
 
 void InfraNetif::CountAddresses(otSysInfraNetIfAddressCounters &aAddressCounters) const
